@@ -4,31 +4,27 @@
   "description": "",
   "author": "FaureWu <502556093@qq.com>",
   "license": "MIT",
-  "main": "react.config.js",
+  "main": "react.config.cjs",
   "files": [
-    "cli.config.js",
-    "eslint.config.js",
-    "react.config.js",
-    "react-base.config.js",
-    "react-typescript.config.js"
+    "cli.config.cjs",
+    "eslint.config.cjs",
+    "react.config.cjs",
+    "typescript.config.cjs"
   ],
   "publishConfig": {
     "access": "public"
   },
   "dependencies": {
-    "@rushstack/eslint-patch": "1.5.1",
-    "@typescript-eslint/eslint-plugin": "6.10.0",
+    "@rushstack/eslint-patch": "1.10.3",
+    "@typescript-eslint/eslint-plugin": "6.21.0",
     "@typescript-eslint/parser": "6.10.0",
-    "eslint": "8.53.0",
-    "eslint-config-prettier": "9.0.0",
-    "eslint-plugin-import": "2.29.0",
+    "eslint": "8.57.0",
+    "eslint-config-prettier": "9.1.0",
+    "eslint-config-standard": "17.1.0",
     "eslint-plugin-jsx-a11y": "6.8.0",
-    "eslint-plugin-n": "16.2.0",
-    "eslint-plugin-prettier": "5.0.1",
-    "eslint-plugin-promise": "6.1.1",
+    "eslint-plugin-prettier": "5.1.3",
     "eslint-plugin-react": "7.33.2",
-    "eslint-plugin-react-hooks": "4.6.0",
-    "prettier": "3.0.3",
-    "typescript": "5.2.2"
+    "eslint-plugin-react-hooks": "4.6.2",
+    "prettier": "3.3.1"
   }
 }
