@@ -1,58 +1,62 @@
-#!/usr/bin/env node
-'use strict'
+import { createRequire } from 'node:module'
+import { Command, Option, InvalidArgumentError } from 'commander'
+import figlet from 'figlet'
 
-const minimist = require('minimist')
-const chalk = require('chalk')
-const logger = require('@doerjs/utils/logger')
+import { cliPackageJsonPath } from '../lib/cliPath.js'
+import create, { validateName } from '../scripts/create.js'
+import dev from '../scripts/dev.js'
+import build from '../scripts/build.js'
 
-const helper = require('../lib/helper')
-process.on('unhandledRejection', (err) => {
-  throw err
-})
+const require = createRequire(import.meta.url)
+const cliPackage = require(cliPackageJsonPath)
 
-helper.logo()
-helper.version()
+console.info(figlet.textSync('Doer', 'Ghost'))
+console.info(`👣 Doer v${cliPackage.version}`)
+console.info()
 
-const argv = minimist(process.argv.slice(2), {
-  string: [],
-  boolean: ['help', 'version'],
-  alias: {
-    version: ['v'],
-    help: ['h'],
-  },
-})
-
-const [command, name] = argv._
-
-const VALID_COMMANDS = ['create', 'dev', 'build']
-const isValidCommand = command && VALID_COMMANDS.some((c) => c === command)
-
-if (!command) {
-  if (argv.version) {
-    // no action
-  } else {
-    helper.help()
+function ensureName(value) {
+  const message = validateName(value)
+  if (message) {
+    throw new InvalidArgumentError(message)
   }
 
-  process.exit(-1)
+  return value
 }
 
-if (!isValidCommand) {
-  logger.fail(
-    `无效的命令参数111 ${chalk.bold(command)}，执行 \`${chalk.bold('doer --help | doer -h')}\` 获取帮助信息。`,
+const cli = new Command()
+
+cli
+  .name('doer')
+  .description('一款集项目模版，研发和打包一体化的工具集合')
+  .version(cliPackage.version, '-v --version', '查看工具版本号')
+  .helpOption('-h, --help', '查看帮助信息')
+  .helpCommand(false)
+
+cli
+  .command('create')
+  .description('创建模版项目')
+  .argument('[name]', '项目或者库名称', ensureName)
+  .addOption(
+    new Option('-t, --type <type>', '创建的模版类型，project项目类型，library库类型').choices(['project', 'library']),
   )
-  console.info()
-  process.exit(-1)
-}
+  .addOption(new Option('-s, --style <style>', '使用哪种样式处理器').choices(['css', 'less']))
+  .option('--typescript', '是否使用typescript')
+  .action((name, options) => {
+    create({ name, ...options })
+  })
 
-switch (command) {
-  case 'create':
-    require('../scripts/create')({ name })
-    break
-  case 'dev':
-    require('../scripts/dev')()
-    break
-  case 'build':
-    require('../scripts/build')()
-    break
-}
+cli
+  .command('dev')
+  .description('启动开发环境')
+  .action(() => {
+    dev()
+  })
+
+cli
+  .command('build')
+  .description('打包项目工程')
+  .action(() => {
+    build()
+  })
+
+cli.parse()
