@@ -1,12 +1,9 @@
-'use strict'
+import path from 'node:path'
+import * as env from '@doerjs/utils/env.js'
 
-const path = require('node:path')
-const is = require('@doerjs/utils/is')
-const env = require('@doerjs/utils/env')
+import Mock from './Mock.js'
 
-const Mock = require('./Mock')
-
-module.exports = function (plugin, option) {
+export default function (plugin) {
   const mock = new Mock()
 
   function createMock(devServer) {
@@ -15,28 +12,20 @@ module.exports = function (plugin, option) {
     }
   }
 
-  plugin.hooks.environment.tap('Mock', (context) => {
+  plugin.hooks.context.tap((context) => {
     env.setBoolean('MOCK', false)
     env.setNumber('MOCK_DELAY', 0)
     env.setString('MOCK_SERVER_PREFIX', '/mock')
 
-    mock.setMockPath(path.resolve(context.paths.runtimePath, 'mocks'))
+    mock.setMockPath(path.resolve(context.path.runtime, 'mocks'))
   })
 
-  plugin.hooks.devServer.tap('Mock', (config) => {
-    if (is.isFunction(config.setupMiddlewares)) {
-      const oldSetupMiddlewares = config.setupMiddlewares
-      config.setupMiddlewares = function (middlewares, devServer) {
-        createMock(devServer)
-        const nextMiddlewares = oldSetupMiddlewares(middlewares, devServer)
-        return nextMiddlewares
-      }
-      return
-    }
-
-    config.setupMiddlewares = function (middlewares, devServer) {
+  plugin.hooks.webpackDevServerConfigure.tap((webpackDevServerConfigure) => {
+    const oldSetupMiddlewares = webpackDevServerConfigure.get('setupMiddlewares')?.toValue()
+    webpackDevServerConfigure.set('setupMiddlewares', function (middlewares, devServer) {
       createMock(devServer)
-      return middlewares
-    }
+      const nextMiddlewares = oldSetupMiddlewares ? oldSetupMiddlewares(middlewares, devServer) : middlewares
+      return nextMiddlewares
+    })
   })
 }
