@@ -1,68 +1,57 @@
-'use strict'
-
 function check(data, type) {
   return Object.prototype.toString.call(data) === `[object ${type}]`
 }
 
-function isUndefined(data) {
+export function isUndefined(data) {
   return check(data, 'Undefined')
 }
 
-function isNull(data) {
+export function isNull(data) {
   return check(data, 'Null')
 }
 
-function isString(data) {
+export function isString(data) {
   return check(data, 'String')
 }
 
-function isFunction(data) {
+export function isFunction(data) {
   return check(data, 'Function')
 }
 
-function isObject(data) {
+export function isObject(data) {
   return check(data, 'Object')
 }
 
-function isNumber(data) {
+// 是否是字面量对象
+export function isLiteObject(data) {
+  return data && data.constructor === Object
+}
+
+export function isNumber(data) {
   return check(data, 'Number')
 }
 
-function isRegExp(data) {
+export function isRegExp(data) {
   return check(data, 'RegExp')
 }
 
-function isArray(data) {
+export function isArray(data) {
   return check(data, 'Array')
 }
 
-function isDate(data) {
+export function isDate(data) {
   return check(data, 'Date')
 }
 
-function isBoolean(data) {
+export function isBoolean(data) {
   return check(data, 'Boolean')
 }
 
-function isEnum(enums = []) {
+export function isEnum(enums = []) {
   return function (data) {
     if (isArray(enums)) {
       return enums.some((item) => item === data)
     }
     return false
   }
-}
-
-module.exports = {
-  isUndefined,
-  isNull,
-  isString,
-  isFunction,
-  isObject,
-  isNumber,
-  isRegExp,
-  isArray,
-  isDate,
-  isBoolean,
-  isEnum,
 }
